Move abstract fixture to module scope
